+++ src/test/groovy/io/github/devatherock/test/ArtifactoryControllerSpec.groovy
@@ -1,0 +1,48 @@
+package io.github.devatherock.test
+
+import javax.inject.Inject
+
+import io.micronaut.http.HttpRequest
+import io.micronaut.http.client.HttpClient
+import io.micronaut.http.client.annotation.Client
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
+import spock.lang.Specification
+import spock.lang.Unroll
+
+/**
+ * Test class for {@link ArtifactoryController}
+ */
+@MicronautTest
+class ArtifactoryControllerSpec extends Specification {
+
+    @Inject
+    @Client('${test.server.url}')
+    HttpClient httpClient
+
+    void 'test get file content'() {
+        when:
+        String manifest = httpClient.toBlocking().retrieve(
+                HttpRequest.GET('/artifactory/manifest.json')
+                        .header('X-JFrog-Art-Api', 'dummy'))
+
+        then:
+        manifest.contains('layers')
+    }
+
+    @Unroll
+    void 'test get folder info - #path'() {
+        when:
+        String folderInfo = httpClient.toBlocking().retrieve(
+                HttpRequest.GET("/artifactory/${path}")
+                        .header('X-JFrog-Art-Api', 'dummy'))
+
+        then:
+        folderInfo.contains(response)
+
+        where:
+        path                        | response
+        'api/storage/manifest.json' | 'downloadCount'
+        'api/storage/latest'        | 'lastModified'
+        'api/storage/notfound'      | 'uri'
+    }
+}